# functions/api-specs/v1/openapi.yml
openapi: 3.0.3
info:
  title: Whatboard API
  description: |
    ## Introduction

    The Whatboard API allows you to use your own (or 3rd-party) software to interact with your content on
    Whatboard. You can:

    - Create Boards
    - Update Boards
    - Delete Boards
    - List Boards
    - Retrieve individual Boards

    - Add Blocks to a Board
    - Update Blocks in a Board
    - Delete Blocks from a Board
    - List Blocks in a Board
    - Retrieve individual Blocks in a Board

    ## Limitations

    - It is not currently possible to invite users to collaborate on a Board using this API
    - It is not currently possible to upload files using this API

    ## Authentication

    The Whatboard API uses OpenID Connect to handle authentication.

    The configuration document is located at the following URL: [https://whatboard.app/oidc/.well-known/openid-configuration](https://whatboard.app/oidc/.well-known/openid-configuration)

    You can learn more about OpenID Connect from these sites:

    - [OpenID Connect Playground](https://openidconnect.net/)
    - [OpenID Specification Site](https://openid.net)

    ### Supported flows (`grant_type`s)

    - `implicit`
    - `authorization_code` (with PKCE support)
    - `refresh_token`

    Browser-based clients are encouraged to use the `authorization_code` flow with PKCE. Only use the
    `implicit` flow if using `authorization_code` with PKCE is not possible.

    ### API Docs Note

...
  version: 1.0.0
...
  contact:
...
servers:
...
security:
  - OIDC: []
...
paths:
...
  "/blocks/{boardId}":
    parameters:
      - $ref: "#/components/parameters/BoardId"
    get:
      summary: Retrieve a list of Blocks for the given Board
      parameters:
        - $ref: "#/components/parameters/OrderBy"
        - $ref: "#/components/parameters/PageToken"
        - $ref: "#/components/parameters/EqualTo"
        - $ref: "#/components/parameters/Dir"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: A list of Blocks
          content:
            application/json:
              schema:
                type: object
                required:
                  - result
                properties:
                  result:
                    title: Result
                    type: array
                    items:
                      $ref: "#/components/schemas/BlockOutput"
                  prevPageToken:
                    title: Previous Page Token
                    description: The pagination token for the previous page of results
                    type: string
                  nextPageToken:
                    title: Next Page Token
                    description: The pagination token for the next page of results
                    type: string
        default:
          $ref: "#/components/responses/GenericError"
      description: Retrieve a list of Blocks for the given Board
      operationId: listBlocks
      tags:
        - blocks
    post:
      summary: Add a Block to the given Board
      requestBody:
        $ref: "#/components/requestBodies/Block"
      responses:
        "200":
          $ref: "#/components/responses/Block"
        default:
          $ref: "#/components/responses/GenericError"
      description: Add a Block to the given Board
      operationId: createBlock
      tags:
        - blocks
  "/blocks/{boardId}/{blockId}":
    parameters:
      - $ref: "#/components/parameters/BoardId"
      - $ref: "#/components/parameters/BlockId"
    get:
      summary: Retrieve a single Block by ID
      responses:
        "200":
          $ref: "#/components/responses/Block"
        default:
          $ref: "#/components/responses/GenericError"
      description: Retrieve a single Block by ID
      operationId: getBlock
      tags:
        - blocks
    patch:
      summary: Update a Block
      requestBody:
        $ref: "#/components/requestBodies/Block"
      responses:
        "200":
          $ref: "#/components/responses/Block"
        default:
          $ref: "#/components/responses/GenericError"
      description: Update a Block
      operationId: updateBlock
      tags:
        - blocks
    delete:
      summary: Delete a Block
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        default:
          $ref: "#/components/responses/GenericError"
      description: Delete a Block
      operationId: deleteBlock
      tags:
        - blocks
...
components:
  schemas:
...
    BlockIdProps:
      title: Block ID Props
      type: object
      properties:
        id:
          title: Block ID
          type: string
        board_id:
          title: Board ID
          type: string
        created_by:
          title: Created By
          type: string
      required:
        - id
        - board_id
        - created_by
    BlockInputProps:
      title: Block Input Props
      type: object
      required:
        - type
      properties:
        title:
          title: Title
          type: string
          default: Title
        type:
          title: Type
          type: string
        color:
          title: Color
          type: string
          enum:
            - "#dc143c"
            - "#FF2400"
            - "#f67c09"
            - "#ffc106"
            - "#55C57C"
            - "#0da00b"
            - "#0f8df0"
            - "#2c387e"
            - "#051225"
            - "#a043bc"
            - "#F4869C"
            - "#FCEAB4"
            - "#795548"
            - "#4B4B4B"
            - "#1f1f1f"
            - "#000000"
            - "#F1F1F1"
            - "#ffffff"
    BlockOutputProps:
      title: Block Output Props
      type: object
      allOf:
        - $ref: "#/components/schemas/Grid"
        - type: object
          required:
            - date_created
          properties:
            date_created:
              title: Date Created
              type: string
              format: date-time
            static:
              title: "Grid: Static Placement"
              description: '"Static" Blocks are neither draggable, nor resizable'
              type: boolean
              default: false
            minH:
              title: "Grid: Minimum Height"
              description: 'The minimum height of the Block, in "grid units"'
              type: integer
              default: 1
            minW:
              title: "Grid: Minimum Width"
              description: 'The minimum width of the Block, in "grid units"'
              type: integer
              default: 1
            grid:
              title: "Grid: Responsive Layout"
              description: Layout overrides for various screen sizes
              type: object
              properties:
                lg:
                  $ref: "#/components/schemas/Grid"
                md:
                  $ref: "#/components/schemas/Grid"
                sm:
                  $ref: "#/components/schemas/Grid"
                xs:
                  $ref: "#/components/schemas/Grid"
                xxs:
                  $ref: "#/components/schemas/Grid"
            metadata:
              title: Metadata
              type: object
              properties:
                lastModified:
                  title: Last Modified
                  type: string
                  format: date-time
                lastViewed:
                  title: Last Viewed
                  type: object
                  additionalProperties:
                    x-additionalPropertiesName: User WBID
                    type: string
                    format: date-time
    BlockInput:
      title: Block (input)
      type: object
      oneOf:
        - $ref: "#/components/schemas/TextBlockInput"
        - $ref: "#/components/schemas/PDFBlockInput"
        - $ref: "#/components/schemas/ImageBlockInput"
        - $ref: "#/components/schemas/VideoBlockInput"
        - $ref: "#/components/schemas/ButtonsBlockInput"
        - $ref: "#/components/schemas/ChecklistBlockInput"
        - $ref: "#/components/schemas/ConversationBlockInput"
        - $ref: "#/components/schemas/ScriptEmbedBlockInput"
        - $ref: "#/components/schemas/IframeBlockInput"
        - $ref: "#/components/schemas/FilesBlockInput"
        - $ref: "#/components/schemas/QAFormBlockInput"
        - $ref: "#/components/schemas/RSSBlockInput"
        - $ref: "#/components/schemas/FileRequestBlockInput"
        - $ref: "#/components/schemas/GridBlockInput"
      discriminator:
        propertyName: type
        x-explicitMappingOnly: true
        mapping:
          Text: "#/components/schemas/TextBlockInput"
          PDF: "#/components/schemas/PDFBlockInput"
          Image: "#/components/schemas/ImageBlockInput"
          Video: "#/components/schemas/VideoBlockInput"
          Buttons: "#/components/schemas/ButtonsBlockInput"
          Checklist: "#/components/schemas/ChecklistBlockInput"
          Conversation: "#/components/schemas/ConversationBlockInput"
          ScriptEmbed: "#/components/schemas/ScriptEmbedBlockInput"
          Iframe: "#/components/schemas/IframeBlockInput"
          Files: "#/components/schemas/FilesBlockInput"
          QAForm: "#/components/schemas/QAFormBlockInput"
          RSS: "#/components/schemas/RSSBlockInput"
          FileRequest: "#/components/schemas/FileRequestBlockInput"
          Grid: "#/components/schemas/GridBlockInput"
    BlockOutput:
      title: Block (output)
      type: object
      oneOf:
        - $ref: "#/components/schemas/TextBlockOutput"
        - $ref: "#/components/schemas/PDFBlockOutput"
        - $ref: "#/components/schemas/ImageBlockOutput"
        - $ref: "#/components/schemas/VideoBlockOutput"
        - $ref: "#/components/schemas/ButtonsBlockOutput"
        - $ref: "#/components/schemas/ChecklistBlockOutput"
        - $ref: "#/components/schemas/ConversationBlockOutput"
        - $ref: "#/components/schemas/ScriptEmbedBlockOutput"
        - $ref: "#/components/schemas/IframeBlockOutput"
        - $ref: "#/components/schemas/FilesBlockOutput"
        - $ref: "#/components/schemas/QAFormBlockOutput"
        - $ref: "#/components/schemas/RSSBlockOutput"
        - $ref: "#/components/schemas/FileRequestBlockOutput"
        - $ref: "#/components/schemas/GridBlockOutput"
      discriminator:
        propertyName: type
        x-explicitMappingOnly: true
        mapping:
          Text: "#/components/schemas/TextBlockOutput"
          PDF: "#/components/schemas/PDFBlockOutput"
          Image: "#/components/schemas/ImageBlockOutput"
          Video: "#/components/schemas/VideoBlockOutput"
          Buttons: "#/components/schemas/ButtonsBlockOutput"
          Checklist: "#/components/schemas/ChecklistBlockOutput"
          Conversation: "#/components/schemas/ConversationBlockOutput"
          ScriptEmbed: "#/components/schemas/ScriptEmbedBlockOutput"
          Iframe: "#/components/schemas/IframeBlockOutput"
          Files: "#/components/schemas/FilesBlockOutput"
          QAForm: "#/components/schemas/QAFormBlockOutput"
          RSS: "#/components/schemas/RSSBlockOutput"
          FileRequest: "#/components/schemas/FileRequestBlockOutput"
          Grid: "#/components/schemas/GridBlockOutput"
    TextBlockCommon:
      title: Text (common)
      type: object
      properties:
        type:
          type: string
          enum:
            - Text
        text:
          title: Text
          type: string
        isPrivateTextBlock:
          title: Make Private
          type: boolean
          default: false
        isTransparentTextBlock:
          title: Transparent Background
          type: boolean
          default: false
        textEditorFiles:
          title: Text Editor Files
          description: Represents an image uploaded to Whatboard file storage and used within the text editor
          type: array
          items:
            $ref: "#/components/schemas/ImageFile"
    TextBlockInput:
      title: Text (input)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/TextBlockCommon"
    TextBlockOutput:
      title: Text (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/TextBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
    PDFBlockCommon:
      title: PDF (common)
      type: object
      properties:
        type:
          type: string
          enum:
            - PDF
        pdfFiles:
          title: PDF Files
          description: Represents a PDF uploaded to Whatboard file storage
          type: array
          maxItems: 1
          items:
            $ref: "#/components/schemas/PDFFile"
        pdf_path:
          title: PDF URL
          description: "The URL to a PDF file. Typically, this will be a PDF file hosted outside of Whatboard. However, if the `pdfFiles` property is present, then this property will automatically be set using data from `pdfFiles`"
          type: string
          format: url
...
    PDFBlockOutput:
      title: PDF (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/PDFBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
    ImageBlockCommon:
      title: Image (common)
      type: object
      properties:
        type:
          title: Type
          type: string
          enum:
            - Image
        imageFiles:
          title: Image Files
          description: Represents an image uploaded to Whatboard file storage
          type: array
          maxItems: 1
          items:
            $ref: "#/components/schemas/ImageFile"
        image_path:
          title: Image URL
          description: "The URL to an image file. Typically, this will be an image file hosted outside of Whatboard. However, if the `imageFiles` property is present, then this property will automatically be set using data from `imageFiles`"
          type: string
          format: url
        imageLink:
          title: Image Link
          description: When the image is clicked on, this web page will be opened in a new browser window/tab.
          type: string
          format: url
...
    ImageBlockOutput:
      title: Image (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/ImageBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
    VideoBlockCommon:
      title: Video (common)
      type: object
      properties:
        type:
          type: string
          enum:
            - Video
        video_id:
          title: Video URL
          type: string
          format: url
...
    VideoBlockOutput:
      title: Video (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/VideoBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
    ButtonsBlockCommon:
      title: Buttons (common)
      type: object
      allOf:
        - type: object
          properties:
            type:
              type: string
              enum:
                - Buttons
        - $ref: "#/components/schemas/SupportsWebhooks"
        - type: object
          properties:
            useButtonWebhook:
              title: Use Webhook
              type: boolean
              default: false
            buttons:
              title: Buttons
              type: array
              items:
                $ref: "#/components/schemas/Button"
...
    ButtonsBlockOutput:
      title: Buttons (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/ButtonsBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
    ChecklistBlockCommon:
      title: Checklist (common)
      type: object
      properties:
        type:
          type: string
          enum:
            - Checklist
        data:
          title: Checklist Items
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"
...
    ChecklistBlockOutput:
      title: Checklist (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/ChecklistBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
    ConversationBlockCommon:
      title: Chat (common)
      type: object
      allOf:
        - type: object
          properties:
            type:
              title: Type
              type: string
              enum:
                - Conversation
        - $ref: "#/components/schemas/SupportsWebhooks"
        - type: object
          properties:
            useWebhook:
              title: Use Webhook
              type: boolean
              default: false
            isFireMyActions:
              title: Fire on my actions
              type: boolean
              default: false
...
    ConversationBlockOutput:
      title: Chat (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/ConversationBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
    ScriptEmbedBlockCommon:
      title: Embed HTML (common)
      type: object
      properties:
        type:
          type: string
          enum:
            - ScriptEmbed
        scriptEmbed:
          title: Script Embed Code
          type: string
...
    ScriptEmbedBlockOutput:
      title: Embed HTML (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/ScriptEmbedBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
    IframeBlockCommon:
      title: Embed Web Page (common)
      type: object
      properties:
        type:
          type: string
          enum:
            - Iframe
        iframe_url:
          title: Web Page URL
          type: string
          format: url
        useIframely:
          title: "Use Iframely (ideal for Twitter, Facebook & hard-to-embed sites)"
          type: boolean
          default: false
...
    IframeBlockOutput:
      title: Embed Web Page (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/IframeBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
    FilesBlockCommon:
      title: File Storage (common)
      type: object
      properties:
        type:
          type: string
          enum:
            - Files
        files:
          title: Files
          description: Represents files uploaded to Whatboard file storage
          type: array
          items:
            $ref: "#/components/schemas/File"
...
    FilesBlockOutput:
      title: File Storage (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/FilesBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
    QAFormBlockCommon:
      title: Q & A Form (common)
      type: object
      allOf:
        - type: object
          properties:
            type:
              type: string
              enum:
                - QAForm
        - $ref: "#/components/schemas/SupportsWebhooks"
        - type: object
          properties:
            useQAWebhook:
              title: Use Webhook
              type: boolean
              default: false
            qa_data:
              title: Q & A Form Items
              items:
                $ref: "#/components/schemas/QAFormItem"
              uniqueItems: true
              minItems: 1
            qaEmailBoardOwner:
              title: Email Responses to Board Owner
              type: boolean
              default: true
...
    QAFormBlockOutput:
      title: Q & A Form (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/QAFormBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
    RSSBlockCommon:
      title: RSS Feed (common)
      type: object
      properties:
        type:
          type: string
          enum:
            - RSS
        rss_url:
          title: RSS Feed URL
          type: string
          format: url
...
    RSSBlockOutput:
      title: RSS Feed (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/RSSBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
...
    FileRequestBlockOutput:
      title: File Request (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/FileRequestBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
...
    GridBlockOutput:
      title: Excel / CSV Viewer (output)
      type: object
      allOf:
        - $ref: "#/components/schemas/BlockIdProps"
        - $ref: "#/components/schemas/BlockInputProps"
        - $ref: "#/components/schemas/GridBlockCommon"
        - $ref: "#/components/schemas/BlockOutputProps"
...
    Error:
      title: Error
      type: object
      required:
        - error
      properties:
        error:
          title: Error Message
          type: string
        errorDescription:
          title: Error Description
          type: string
  parameters:
    BoardId:
      name: boardId
      in: path
      schema:
        type: string
      required: true
    BlockId:
      name: blockId
      in: path
      schema:
        type: string
      required: true
    OrderBy:
      name: orderBy
      in: query
      schema:
        type: string
    PageToken:
      name: pageToken
      in: query
      schema:
        type: string
    EqualTo:
      name: equalTo
      in: query
      schema:
        type: string
    Dir:
      name: dir
      in: query
      schema:
        type: string
        enum:
          - asc
          - desc
        default: asc
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 10
  requestBodies:
...
    Block:
      description: A single Block
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BlockInput"
  responses:
...
    Block:
      description: A single Block
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/BlockOutput"
...
    NoContent:
      description: No content
    GenericError:
      description: Generic Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  securitySchemes:
    OIDC:
      type: openIdConnect
      openIdConnectUrl: "https://whatboard.app/oidc/.well-known/openid-configuration"